wasm-parse: refactor
