+++ kernel-ewasm/validator/src/lib.rs
@@ -1,121 +1,137 @@
+//! # Validator
+//!
+//! Module for parsing and validating pwasm contracts on-chain.
 #![cfg_attr(not(feature = "std"), no_std)]
 
 #[cfg(not(feature = "std"))]
 #[macro_use]
 extern crate alloc;
 
-use pwasm_std;
-use pwasm_std::vec::Vec;
-use pwasm_std::String;
-
-pub mod func;
-pub mod import_entry;
-pub mod instructions;
-pub mod io;
+mod func;
+mod import_entry;
+mod instructions;
+mod io;
 mod primitives;
-pub mod serialization;
-pub mod types;
-pub use self::io::Error;
-pub use self::serialization::Deserialize;
+mod serialization;
+mod types;
+use self::serialization::Deserialize;
 
-pub use self::primitives::{
-    CountedList, Uint32, Uint64, Uint8, VarInt32, VarInt64, VarInt7, VarUint1, VarUint32,
-    VarUint64, VarUint7,
+use self::primitives::{
+    CountedList, Uint32, Uint64, Uint8, VarInt32, VarInt64, VarInt7, VarUint1, VarUint32, VarUint7,
 };
 
-/// As per the wasm spec:
-///
-/// Function Index Space
-///
-/// The function index space indexes all imported and internally-defined
-/// function definitions, assigning monotonically-increasing indices based on
-/// the order of definition in the module (as defined by the binary encoding).
-/// Thus, the index space starts at zero with the function imports (if any)
-/// followed by the functions defined within the module.
-// fn get_function_indices() {
-//     // First get the imports
-//     // Second get the functions
-// }
-
-/// A listing is a category of import. There are 3 types of imports whitelisted,
-/// greylisted, and blacklisted. There is no blacklist, everything that is not
-/// whitlisted or greylisted is blacklisted, even if we don't recognise it.
-///
-///  * Whitelisted: Functions which can be run with no state effects and we
-///      don't care about them. Examples include getting addresses, returning,
-///      reverting etc.
-///  * Greylisted: Functions that _do_ perform dangerous operations, but that we
-///      need for the operation of syscalls etc. These calls need to be
-///      surrounded by the correct protections. These are permitted to be
-///      imported, but must be checked for safety.
-///  * Blacklisted: Everything else. These cannot even be imported. If they are
-///      imported the contract is not valid.
-#[derive(Debug)]
-pub enum Listing {
-    White,
-    Grey,
-    Black,
-}
-
-pub trait Listed {
-    fn listing(&self) -> Listing;
-}
-
-#[derive(Debug, Clone)]
-pub struct ImportEntry {
-    index: u32,
-    mod_name: String,
-    field_name: String,
-}
-
-impl Listed for ImportEntry {
-    fn listing(&self) -> Listing {
-        // Nothing should need to be imported from outside "env", but let's
-        // blacklist it just in case.
-        if self.mod_name != "env" {
-            Listing::Black
-        } else {
-            // Tehcnically we don't have to list blacklisted items here, but we
-            // do just for clarity.
-            match self.field_name.as_ref() {
-                "memory" => Listing::White,
-                "storage_read" => Listing::White,
-                "storage_write" => Listing::Black,
-                "ret" => Listing::White,
-                "gas" => Listing::White,
-                "input_length" => Listing::White,
-                "fetch_input" => Listing::White,
-                "panic" => Listing::White,
-                "debug" => Listing::White,
-                "ccall" => Listing::Black,
-                "dcall" => Listing::Grey,
-                "scall" => Listing::White,
-                "value" => Listing::White,
-                "create" => Listing::Black,
-                "suicide" => Listing::White,
-                "blockhash" => Listing::White,
-                "blocknumber" => Listing::White,
-                "coinbase" => Listing::White,
-                "difficulty" => Listing::White,
-                "gaslimit" => Listing::White,
-                "timestamp" => Listing::White,
-                "address" => Listing::White,
-                "sender" => Listing::White,
-                "origin" => Listing::White,
-                "elog" => Listing::Black,
-                "extcodesize" => Listing::White,
-                "extcodecopy" => Listing::White,
-                "create2" => Listing::Black,
-                "gasleft" => Listing::White,
-                _ => Listing::Black,
-            }
-        }
-    }
-}
+mod listing;
+pub mod modules;
+pub use modules::Module;
+pub use modules::Function;
+use listing::*;
 
 /// Be able to determine a contracts validity.
 pub trait Validity {
+    /// Tests the object for validity.
     fn is_valid(&self) -> bool;
+}
+
+impl<'a> Validity for modules::Module<'a> {
+    fn is_valid(&self) -> bool {
+        // Now that we have our hooks into the module, let's iterate over the
+        // imports to determine white/grey/black listings. We need to remember
+        // where the function and code data starts.
+
+        // There is only one greylisted item (dcall) so we will just reserve a
+        // place for that rather than maintain a list.
+        let mut dcall_index: Option<usize> = None;
+        let mut gasleft_index: Option<usize> = None;
+        let mut sender_index: Option<usize> = None;
+        if let Some(imports) = self.imports() {
+            for (index, import) in imports.enumerate() {
+                if import.mod_name == "env" && import.field_name == "sender" {
+                    if sender_index.is_some() {
+                        panic!("sender imported multiple times");
+                    }
+                    sender_index = Some(index as usize);
+                }
+
+                if import.mod_name == "env" && import.field_name == "gasleft" {
+                    if gasleft_index.is_some() {
+                        panic!("gasleft imported multiple times");
+                    }
+                    gasleft_index = Some(index as usize);
+                }
+
+                // println!("mod_name: {}, field_name: {}, f_index: {}, listing: {:?}",
+                // import.mod_name, import.field_name, import.index, import.listing());
+                match import.listing() {
+                    Listing::White => (),
+                    Listing::Grey => {
+                        if dcall_index.is_some() {
+                            panic!("dcall imported multiple times");
+                        }
+                        // Document here why this is the case
+                        dcall_index = Some(index as usize);
+                    }
+                    Listing::Black => {
+                        // If we encounter a blacklisted import we can return
+                        // early.
+                        // println!("{:?} is blacklisted", import);
+                        return false;
+                    }
+                }
+            }
+        } else {
+            // println!("no imports",);
+        }
+
+        if let Some(funcs) = self.functions() {
+            for (_i, func) in funcs.enumerate() {
+                // println!("func[{}]: {:?}", i, func);
+                if let (Some(dcall_i), Some(gasleft_i), Some(sender_i)) =
+                    (dcall_index, gasleft_index, sender_index)
+                {
+                    if func.is_syscall(dcall_i as u32, gasleft_i as u32, sender_i as u32) {
+                        // If the function is a system call we can continue past
+                        // it
+                        continue;
+                    }
+                    // At this point we know that the function is not a syscall.
+                    // We must now check that it has no black or grey listed
+                    // calls. We only care about calls here. We only need to do
+                    // this if dcall is imported in the first place.
+                    if let Some(dcall_i) = dcall_index {
+                        if func.contains_grey_call(dcall_i as u32) {
+                            // This function contains a greylisted call (i.e.
+                            // dcall), so we must return with false as the
+                            // contract is invalid.
+                            return false;
+                        }
+                    }
+                }
+            }
+        }
+        // All the tests have passed so we can return true.
+        true
+    }
+}
+
+fn parse_varuint_32(cursor: &mut Cursor) -> u32 {
+    let mut res = 0;
+    let mut shift = 0;
+    loop {
+        if shift > 31 {
+            panic!("invalid varuint32");
+        }
+
+        let b = cursor.read_ref().clone() as u32;
+        res |= (b & 0x7f).checked_shl(shift).expect("invalid varuint32");
+        shift += 7;
+        if (b >> 7) == 0 {
+            if shift >= 32 && (b as u8).leading_zeros() < 4 {
+                panic!("invalid varuint32");
+            }
+            break;
+        }
+    }
+    res
 }
 
 // Seek does not seem to be implemented in core, so we'll reimplement what we
@@ -164,455 +180,10 @@ impl<'a> io::Read for Cursor<'a> {
     }
 }
 
-impl Validity for &[u8] {
-    fn is_valid(&self) -> bool {
-        // Set an index value, which is our offset into the wasm bytes.
-        let mut cursor = Cursor {
-            current_offset: 0,
-            body: self,
-        };
-
-        // The first two steps are to take the magic number and version to check
-        // that it is valid wasm. This is not strictly necessary, as it is the
-        // job of the runtime to ensure the wasm is valid (ad we rely on that
-        // fact), however, it's cheap and allows us prevent future versions of
-        // wasm code being deployed (for which our assumptions may not hold).
-
-        // Take the magic number, check that it matches
-        if cursor.read_ref_n(4) != &[0, 97, 115, 109] {
-            panic!("magic number not found");
-        }
-
-        // Take the version, check that it matches
-        if cursor.read_ref_n(4) != &[1, 0, 0, 0] {
-            panic!("proper version number not found");
-        }
-
-        // Now we should be at the first section. We care about 4 sections:
-        // types, imports, functions, and code. The first thing we want to do is
-        // to find the offsets of these 4 sections. We assume the wasm is well
-        // formed and there are no duplicate sections and the like. It is also
-        // possible some of these sections don't exist.
-        let mut type_section_offset: Option<usize> = None;
-        let mut import_section_offset: Option<usize> = None;
-        let mut function_section_offset: Option<usize> = None;
-        let mut code_section_offset: Option<usize> = None;
-        while cursor.current_offset < self.len() {
-            let section: Section = parse_section(&mut cursor);
-            // There are many section types we don't care about, for example,
-            // Custom sections generally contain debugging symbols and
-            // meaningful function names which are irrelevant to the current
-            // process. We care only about types, imports, functions, and code.
-            match section.type_ {
-                SectionType::Type => {
-                    if type_section_offset.is_some() {
-                        panic!("multiple type sections");
-                    }
-                    type_section_offset = Some(section.offset);
-                }
-                SectionType::Import => {
-                    if import_section_offset.is_some() {
-                        panic!("multiple import sections");
-                    }
-                    import_section_offset = Some(section.offset);
-                }
-                SectionType::Function => {
-                    if function_section_offset.is_some() {
-                        panic!("multiple function sections");
-                    }
-                    function_section_offset = Some(section.offset);
-                }
-                SectionType::Code => {
-                    if code_section_offset.is_some() {
-                        panic!("multiple code sections");
-                    }
-                    code_section_offset = Some(section.offset);
-                }
-                // We ignore any section we are not interested in.
-                _ => (),
-            }
-        }
-        if cursor.current_offset != self.len() {
-            panic!("mismatched length");
-        }
-
-        // Now that we have our hooks into the module, let's iterate over the
-        // imports to determine white/grey/black listings. We need to remember
-        // where the function and code data starts.
-
-        // There is only one greylisted item (dcall) so we will just reserve a
-        // place for that rather than maintain a list.
-        let mut dcall_index: Option<usize> = None;
-        let mut gasleft_index: Option<usize> = None;
-        let mut sender_index: Option<usize> = None;
-        if let Some(imports_offset) = import_section_offset {
-            // Make a new cursor for imports
-            let mut imports_cursor = Cursor {
-                current_offset: imports_offset,
-                body: &self,
-            };
-            let _section_size = parse_varuint_32(&mut imports_cursor);
-            // How many imports do we have?
-            let n_imports = parse_varuint_32(&mut imports_cursor);
-            // println!("n_imports: {}", n_imports);
-            for i in 0..n_imports {
-                // let mut cursor = Cursor {i:0};
-
-                // Here we parse the names of the import, and its function
-                // index.
-                let import = parse_import(&mut imports_cursor, i);
-
-                if import.mod_name == "env" && import.field_name == "sender" {
-                    if sender_index.is_some() {
-                        panic!("sender imported multiple times");
-                    }
-                    sender_index = Some(import.index as usize);
-                }
-
-                if import.mod_name == "env" && import.field_name == "gasleft" {
-                    if gasleft_index.is_some() {
-                        panic!("gasleft imported multiple times");
-                    }
-                    gasleft_index = Some(import.index as usize);
-                }
-
-                // println!("mod_name: {}, field_name: {}, f_index: {}, listing: {:?}",
-                // import.mod_name, import.field_name, import.index, import.listing());
-                match import.listing() {
-                    Listing::White => (),
-                    Listing::Grey => {
-                        if dcall_index.is_some() {
-                            panic!("dcall imported multiple times");
-                        }
-                        // Document here why this is the case
-                        dcall_index = Some(import.index as usize);
-                    }
-                    Listing::Black => {
-                        // If we encounter a blacklisted import we can return
-                        // early.
-                        // println!("{:?} is blacklisted", import);
-                        return false;
-                    }
-                }
-            }
-        }
-
-        // The functions index into types. In fact the function section is just
-        // a vector of type ids. We don't care about types at this stage.
-        if let (Some(functions_offset), Some(code_offset)) =
-            (function_section_offset, code_section_offset)
-        {
-            // Make a new cursor for functions
-            let mut functions_cursor = Cursor {
-                current_offset: functions_offset,
-                body: &self,
-            };
-            // Make a new cursor for code
-            let mut code_cursor = Cursor {
-                current_offset: code_offset,
-                body: &self,
-            };
-            // We will have to try and update these in parallel
-            let _function_section_size = parse_varuint_32(&mut functions_cursor);
-            let _code_section_size = parse_varuint_32(&mut code_cursor);
-            // println!("functions_offset: {:?}", functions_offset);
-            // println!("code_offset: {:?}", code_offset);
-            let n_functions = parse_varuint_32(&mut functions_cursor);
-            let n_bodies = parse_varuint_32(&mut code_cursor);
-
-            // println!("functions_size: {:?}", _function_section_size);
-            // println!("code_size: {:?}", _code_section_size);
-
-            assert_eq!(n_functions, n_bodies);
-
-            // Next we iterate through the function bodies and check if they
-            // violate any of our rules.
-
-            // We now know the location of dcall, if there is one.
-            // We need to iterate over every function and read its code. A
-            // function can be one of three things:
-            //
-            //     * A syscall that follows the format
-            //     * A function which is not a syscall and features a greylisted call.
-            //     * A function which does not contain a greylisted call or a blacklistd call.
-            // The possiblities are checked in that order.
-            for _i in 0..n_bodies {
-                let body_size = parse_varuint_32(&mut code_cursor);
-                // First we check if it is a system call, this is only possible
-                // if we have the three required imports.
-                if let (Some(dcall_i), Some(gasleft_i), Some(sender_i)) =
-                    (dcall_index, gasleft_index, sender_index)
-                {
-                    if is_syscall(
-                        dcall_i as u32,
-                        gasleft_i as u32,
-                        sender_i as u32,
-                        &self[(code_cursor.current_offset)
-                            ..(code_cursor.current_offset + body_size as usize)],
-                    ) {
-                        // If the function is a system call we can continue past it
-                        continue;
-                    }
-                }
-                // At this point we know that the function is not a syscall. We
-                // must now check that it has no black or grey listed calls. We
-                // only care about calls here. We only need to do this if dcall
-                // is imported in the first place.
-                if let Some(dcall_i) = dcall_index {
-                    if contains_grey_call(
-                        dcall_i as u32,
-                        &self[(code_cursor.current_offset)
-                            ..(code_cursor.current_offset + body_size as usize)],
-                    ) {
-                        // This function contains a greylisted call (i.e.
-                        // dcall), so we must return with false as the contract
-                        // is invalid.
-                        return false;
-                    }
-                }
-                // We need to skip over the code body in the main cursor.
-                code_cursor.skip(body_size as usize);
-            }
-        } else {
-            panic!("no code");
-        }
-        // All the tests have passed so we can return true.
-        true
-    }
-}
-
-#[derive(Debug)]
-enum SectionType {
-    Custom,
-    Type,
-    Import,
-    Function,
-    Table,
-    Memory,
-    Global,
-    Export,
-    Start,
-    Element,
-    Code,
-    Data,
-}
-
-#[derive(Debug)]
-struct Section {
-    type_: SectionType,
-    // The offset is the byte offset of the start of this
-    // section, i.e. it points directly to the length byte.
-    offset: usize,
-}
-
-fn parse_section(cursor: &mut Cursor) -> Section {
-    let type_n = cursor.read_ref();
-    let offset = cursor.current_offset;
-    let size_n = parse_varuint_32(cursor);
-    let type_ = n_to_section(type_n);
-    let section = Section { type_, offset };
-    cursor.current_offset += size_n as usize;
-    section
-}
-
-fn n_to_section(byte: &u8) -> SectionType {
-    match byte {
-        0 => SectionType::Custom,
-        1 => SectionType::Type,
-        2 => SectionType::Import,
-        3 => SectionType::Function,
-        4 => SectionType::Table,
-        5 => SectionType::Memory,
-        6 => SectionType::Global,
-        7 => SectionType::Export,
-        8 => SectionType::Start,
-        9 => SectionType::Element,
-        10 => SectionType::Code,
-        11 => SectionType::Data,
-        _ => panic!("invalid section type"),
-    }
-}
-
-fn parse_varuint_32(cursor: &mut Cursor) -> u32 {
-    let mut res = 0;
-    let mut shift = 0;
-    loop {
-        if shift > 31 {
-            panic!("invalid varuint32");
-        }
-
-        let b = cursor.read_ref().clone() as u32;
-        res |= (b & 0x7f).checked_shl(shift).expect("invalid varuint32");
-        shift += 7;
-        if (b >> 7) == 0 {
-            if shift >= 32 && (b as u8).leading_zeros() < 4 {
-                panic!("invalid varuint32");
-            }
-            break;
-        }
-    }
-    res
-}
-
-fn parse_import(cursor: &mut Cursor, index: u32) -> ImportEntry {
-    let mut reader = Cursor {
-        current_offset: cursor.current_offset,
-        body: cursor.body,
-    };
-    let import: import_entry::ImportEntry =
-        import_entry::ImportEntry::deserialize(&mut reader).expect("counted list");
-    let val = ImportEntry {
-        index,
-        mod_name: String::from(import.module()),
-        field_name: String::from(import.field()),
-    };
-    cursor.current_offset = reader.current_offset;
-    val
-}
-
-/// An iterator which counts from one to five
-struct Code<'a> {
-    locals: Vec<func::Local>,
-    current_offset: usize,
-    body: &'a [u8],
-}
-
-// we want our count to start at one, so let's add a new() method to help.
-// This isn't strictly necessary, but is convenient. Note that we start
-// `count` at zero, we'll see why in `next()`'s implementation below.
-impl<'a> Code<'a> {
-    fn new(body: &'a [u8]) -> Code {
-        let mut reader = Cursor {
-            current_offset: 0,
-            body: body,
-        };
-        let locals: Vec<func::Local> = CountedList::<func::Local>::deserialize(&mut reader)
-            .expect("counted list")
-            .into_inner();
-        Code {
-            locals,
-            current_offset: reader.current_offset,
-            body: body,
-        }
-    }
-}
-
-impl<'a> Iterator for Code<'a> {
-    type Item = crate::instructions::Instruction;
-
-    fn next(&mut self) -> Option<Self::Item> {
-        if self.current_offset < self.body.len() {
-            // We need to parse the code into something meaningful
-            let mut reader = Cursor {
-                current_offset: self.current_offset,
-                body: self.body,
-            };
-            let val = Some(
-                crate::instructions::Instruction::deserialize(&mut reader)
-                    .expect("expected valid instruction"),
-            );
-            self.current_offset = reader.current_offset;
-            val
-        } else {
-            None
-        }
-    }
-}
-
-pub fn is_syscall(dcall_i: u32, gasleft_i: u32, sender_i: u32, body: &[u8]) -> bool {
-    let mut code_iter = Code::new(body);
-
-    // Check that no locals are used
-    if code_iter.locals.len() > 0 {
-        return false;
-    }
-
-    // First we need to check that the instructions are correct, that is:
-    //   0. call $a
-    //   1. call $b
-    //   2. get_local 0
-    //   3. get_local 1
-    //   4. get_local 2
-    //   5. get_local 3
-    //   6. call $c
-    // $a, $b, and $c will be used later.
-
-    //   0. call gasleft
-    if let Some(instructions::Instruction::Call(f_ind)) = code_iter.next() {
-        if f_ind != gasleft_i {
-            return false;
-        }
-    } else {
-        return false;
-    }
-    //   1. call sender
-    if let Some(instructions::Instruction::Call(f_ind)) = code_iter.next() {
-        if f_ind != sender_i {
-            return false;
-        }
-    } else {
-        return false;
-    }
-    //   2. get_local 0
-    if let Some(instructions::Instruction::GetLocal(0)) = code_iter.next() {
-    } else {
-        return false;
-    }
-    //   3. get_local 1
-    if let Some(instructions::Instruction::GetLocal(1)) = code_iter.next() {
-    } else {
-        return false;
-    }
-    //   4. get_local 2
-    if let Some(instructions::Instruction::GetLocal(2)) = code_iter.next() {
-    } else {
-        return false;
-    }
-    //   5. get_local 3
-    if let Some(instructions::Instruction::GetLocal(3)) = code_iter.next() {
-    } else {
-        return false;
-    }
-
-    //   6. call dcall
-    if let Some(instructions::Instruction::Call(f_ind)) = code_iter.next() {
-        if f_ind != dcall_i {
-            return false;
-        }
-    } else {
-        return false;
-    }
-    //   7. END
-    if let Some(instructions::Instruction::End) = code_iter.next() {
-    } else {
-        return false;
-    }
-    // We have checked locals and code, we don't really care abou the type, so
-    // we can return true.
-    true
-}
-
-// TODO: we need to account for indirect calls too.
-pub fn contains_grey_call(dcall_i: u32, body: &[u8]) -> bool {
-    let code_iter = Code::new(body);
-    for instruction in code_iter {
-        // We only care about Call instructions
-        if let instructions::Instruction::Call(f_ind) = instruction {
-            // if f_ind is a grey call then we return true, as we are asking the
-            // question "Does this function contain a call to a greylisted
-            // import?".
-            if f_ind == dcall_i {
-                return true;
-            }
-        }
-    }
-    // No instructions were greylisted, so we can return false.
-    false
-}
-
 #[cfg(test)]
 mod tests {
     use super::*;
+    use modules::Module;
     use std::fs::File;
     use std::io::Read;
     use wabt::wat2wasm;
@@ -621,7 +192,7 @@ mod tests {
     fn module_only_pass() {
         let wat = "(module)";
         let wasm = wat2wasm(wat).unwrap();
-        let validation_result = wasm.as_slice().is_valid();
+        let validation_result = Module::new(wasm.as_slice()).is_valid();
         assert_eq!(validation_result, true);
     }
 
@@ -636,7 +207,7 @@ mod tests {
   (export "call" (func $call)))
 "#;
         let wasm = wat2wasm(wat).unwrap();
-        let validation_result = wasm.as_slice().is_valid();
+        let validation_result = Module::new(wasm.as_slice()).is_valid();
         assert_eq!(validation_result, true);
     }
 
@@ -648,8 +219,8 @@ mod tests {
         .expect("could not open file");
         let mut wasm = Vec::new();
         f.read_to_end(&mut wasm).unwrap();
-        let validation_result = wasm.as_slice().is_valid();
-        assert_eq!(validation_result, true);
+        let validation_result = Module::new(wasm.as_slice()).is_valid();
+        assert_eq!(validation_result, false);
     }
 
     #[test]
@@ -659,7 +230,7 @@ mod tests {
         let mut wat = Vec::new();
         f.read_to_end(&mut wat).unwrap();
         let wasm = wat2wasm(wat).unwrap();
-        let validation_result = wasm.as_slice().is_valid();
+        let validation_result = Module::new(wasm.as_slice()).is_valid();
         assert_eq!(validation_result, true);
     }
 
@@ -670,7 +241,7 @@ mod tests {
         let mut wat = Vec::new();
         f.read_to_end(&mut wat).unwrap();
         let wasm = wat2wasm(wat).unwrap();
-        let validation_result = wasm.as_slice().is_valid();
+        let validation_result = Module::new(wasm.as_slice()).is_valid();
         assert_eq!(validation_result, false);
     }
 
@@ -681,7 +252,7 @@ mod tests {
         let mut wat = Vec::new();
         f.read_to_end(&mut wat).unwrap();
         let wasm = wat2wasm(wat).unwrap();
-        let validation_result = wasm.as_slice().is_valid();
+        let validation_result = Module::new(wasm.as_slice()).is_valid();
         assert_eq!(validation_result, false);
     }
 
@@ -692,7 +263,7 @@ mod tests {
         let mut wat = Vec::new();
         f.read_to_end(&mut wat).unwrap();
         let wasm = wat2wasm(wat).unwrap();
-        let validation_result = wasm.as_slice().is_valid();
+        let validation_result = Module::new(wasm.as_slice()).is_valid();
         assert_eq!(validation_result, false);
     }
 
@@ -712,7 +283,7 @@ mod tests {
   (export "call" (func $call)))
 "#;
         let wasm = wat2wasm(wat).unwrap();
-        let validation_result = wasm.as_slice().is_valid();
+        let validation_result = Module::new(wasm.as_slice()).is_valid();
         assert_eq!(validation_result, false);
     }
 }
